Ulicms 2018.2 Release into master

+++ ulicms/content/modules/core_package_manager/templates/do_post_install.php
@@ -4,15 +4,20 @@ if ($acl->hasPermission ( "update_system" )) {
 	?>
 <h1><?php translate("run_post_install_script");?></h1>
 <?php
-	// @FIXME: Prüfen, ob die Datei überhaupt existiert
 	$postinstall = ULICMS_ROOT . "/post-install.php";
-	include $postinstall;
-	unlink ( $postinstall );
-	?>
+	if (file_exists ( $postinstall )) {
+		include $postinstall;
+		unlink ( $postinstall );
+		?>
 		<?php if(!file_exists($postinstall)){?>
 <p><?php translate("finished");?></p>
 <?php } ?>
 <?php
+	} else {
+		?>
+<p><?php translate("file_not_found");?></p>
+<?php
+	}
 } else {
 	noperms ();
 }
